CP exec_all.py executes model on all instances
Results for each instance are saved in out/. Actual solutions are saved into out-ins-X.txt in the required format. Chuffed is being used, which will output solutions that satisfy the constraints even if not optimal.
To discriminate between the optimal and non optimal ones, statistics are dumped into stat-ins-X.json files (in out/, too). Non-optimal solutions are marked as 'SATISFIED'.