Adapted CP model to rotations
We are keeping them separate as the performance are definitely different.

Rotations are managed through a set of indeces, used as variable. If an index is inside the set, that means that the corresponding circuit is going to be considered as rotated.
A new array of circuits functionally defined by said set is then used instead of the input parameters.

Output format is the same, but rotated rectangles are explicitly outputted as rotated (width and height inverted). Output order of said circuits is consistent.

Height boundaries have still not been adapted. They should take into account rotated circuits.

Further sperimentation is needed to find a more suitable heuristic for the set search. Unfortunately (but as expected) this new model is not able to solve (in time) the hardest instances. Set search do not give a lot of freedom, an idea could be channeling it into an array of boolean variables (or substituting it with an array altogether) in order to experiment with heuristics (eg. area based or first fail).